Allow registering metrics with the same name but different label names
The reason for this is that we want to support using prometheus-net as a bridge to report data originating from metrics systems that do not have such limitations (such as the .NET 6 Meters API). Such scenarios can create any combinations of label names. This is permissible by OpenMetrics, though violates the Prometheus client authoring requirements (which is OK - what can you do).